requirements.txt
asgiref==3.3.1
astroid==2.4.1
attrs==20.3.0
base58==2.0.1
background==0.1.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Django==3.1.3
django-background-tasks==1.2.5
django-compat==1.0.15
django-filter==2.4.0
djangorestframework==3.12.2
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.9.5
future==0.18.2
hexbytes==0.2.1
idna==2.10
ipfshttpclient==0.7.0a1
isort==4.3.21
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
lazy-object-proxy==1.4.3
livereload==2.6.2
lunr==0.5.8
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
mkdocs==1.1.2
multiaddr==0.0.9
netaddr==0.8.0
nltk==3.5
parsimonious==0.8.1
protobuf==3.14.0
pylint==2.5.2
pyrsistent==0.17.3
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
rlp==2.0.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
toolz==0.11.1
tornado==6.0.4
tqdm==4.48.2
urllib3==1.26.2
varint==1.0.2
web3==5.13.1
websockets==8.1
wrapt==1.12.1
